% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csawRun.R
\name{csawRun}
\alias{csawRun}
\title{Run csaw with command line parameters}
\usage{
csawRun(cli = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{cli}{The command line arguments specifying how to run csaw. By
default this is the trailing arguments from the script that calls this.}
}
\value{
N/A
}
\description{
This is the main function that runs csaw. Intended to be called from an
executable script and to recieve the command line parameters that script
was called with. For the meaning of arguments, see parseCli().
}
