% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCli.R
\name{extractArgs}
\alias{extractArgs}
\title{Extract the arguments from the parsed cli}
\usage{
extractArgs(parsedCli)
}
\value{
A character vector of the arguments, or NULL if none. Will extract
arguments mixed in with options, but does not recognize "-" as a signal
to stop identifying options and use rest as arguments.
}
\description{
Extract the arguments from the parsed cli
}
\keyword{internal}
