% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCli.R
\name{extractOptions}
\alias{extractOptions}
\title{Extract the options from the parsed cli}
\usage{
extractOptions(parsedCli)
}
\value{
A list of the option names and their values, or NULL if no options.
Will extract options mixed in with arguments, but does not recognize "-" as
a signal to stop identifying options and use rest as arguments.
}
\description{
Extract the options from the parsed cli
}
\keyword{internal}
