% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naDecode}
\alias{naDecode}
\title{Convert string "NA" to NA_character_}
\usage{
naDecode(x, naSymbol = "NA")
}
\arguments{
\item{x}{A character vector with string elements that should be NA.}

\item{naSymbol}{The string to convert to NA when found as an element.}
}
\value{
A character vector with missing values (NA) in place of strings
  encoding them, e.g. "NA"
}
\description{
Convert the "NA" elements in a character vector into missing (NA) values.
}
\examples{
naDecode( c("N", "NA", "NAN") )
#> [1] "N"   NA  "NAN"
naDecode( c("A", "B", ".", "D"), naSymbol= "." )
#> [1] A  B  NA  D

}
