% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCli.R
\name{optionRules}
\alias{optionRules}
\title{Define the options for the cli}
\usage{
optionRules()
}
\value{
A list with each element describing an option as an
  optparse::OptionParserOption object
}
\description{
Define the options for the cli
}
\keyword{internal}
