% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParam.R
\name{readParam}
\alias{readParam}
\title{Create readParam object from options}
\usage{
readParam(opts)
}
\arguments{
\item{opts}{The parsed command line options}
}
\value{
A csaw::readParam object.
}
\description{
Create readParam object from options
}
\keyword{internal}
