% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{stopAs}
\alias{stopAs}
\title{Errors by name.}
\usage{
stopAs(name, ...)
}
\arguments{
\item{name}{The name of the error. If this ends with a number, that is the
number of extra parameters needed to format the error message.}

\item{...}{Extra parameters used to format the error message}
}
\value{
N/A (exits with error)
}
\description{
Errors by name.
}
\keyword{internal}
