% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCli.R
\name{validateCli}
\alias{validateCli}
\title{Validate the options ande arguments from the parsed cli}
\usage{
validateCli(opts, args)
}
\arguments{
\item{opts}{Options as a key=value list.}

\item{args}{Arguments as a character vector, in original order.}
}
\value{
A list of key=value settings from validated options, arguments,
and defaults.
}
\description{
Validate the options ande arguments from the parsed cli
}
\keyword{internal}
