% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCli.R
\name{parseCli}
\alias{parseCli}
\title{Parse options and arguments for running csaw}
\usage{
parseCli(cli = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{cli}{The command line arguments specifying how to run csaw. By
default this is the trailing arguments from a script this is run by.}
}
\value{
The parsed cli object
}
\description{
Parse options and arguments for running csaw
}
